+++ alacritty_terminal/CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## 0.25.1-dev
 
+### Changed
+
+- Pass `-q` to `login` on macOS if `~/.hushlogin` is present
+
 ## 0.25.0
 
 ### Changed